requirements.txt
sentence-transformers==5.3.0
torch>=2.0.0
scikit-learn>=1.6.0
feedparser>=6.0.10
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0
telethon>=1.34.0
vk-api>=11.9.9
python-dotenv>=1.0.0
